Address code review feedback
- Address code review feedback
- Add unit tests for PreviewViewModel
- Remove example test files
