+++ app/build.gradle
@@ -11,7 +11,7 @@ android {
 
     defaultConfig {
         applicationId "com.google.jetpackcamera"
-        minSdk 24
+        minSdk 21
         targetSdk 33
         versionCode 1
         versionName "1.0"
@@ -62,13 +62,13 @@ dependencies {
     debugImplementation 'androidx.compose.ui:ui-tooling'
 
     // Compose - Integration with ViewModels
-    implementation 'androidx.lifecycle:lifecycle-viewmodel-compose:2.5.1'
+    implementation 'androidx.lifecycle:lifecycle-viewmodel-compose'
 
     // Compose - Integration with Activities
-    implementation 'androidx.activity:activity-compose:1.5.1'
+    implementation 'androidx.activity:activity-compose'
 
     // Compose - Testing
-    androidTestImplementation "androidx.compose.ui:ui-test-junit4:$compose_version"
+    androidTestImplementation "androidx.compose.ui:ui-test-junit4"
 
     // Testing
     testImplementation 'junit:junit:4.13.2'
@@ -85,6 +85,7 @@ dependencies {
     // Accompanist - Permissions
     implementation 'com.google.accompanist:accompanist-permissions:0.26.5-rc'
 
+    // Camera Preview
     implementation(project(":feature:preview"))
 }
