+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,6 @@
 <resources>
     <string name="app_name">Jetpack Camera</string>
+    <string name="camera_permission_required_rationale">The camera is important for this app. Please grant the permission.</string>
+    <string name="camera_not_available">Camera not available</string>
+    <string name="request_permission">Request permission</string>
 </resources>
