+++ camera-viewfinder-compose/build.gradle
@@ -1,12 +1,10 @@
 plugins {
     id 'com.android.library'
     id 'org.jetbrains.kotlin.android'
-    id 'kotlin-kapt'
-    id 'com.google.dagger.hilt.android'
 }
 
 android {
-    namespace 'com.google.jetpackcamera.feature.preview'
+    namespace 'com.google.jetpackcamera.camerax'
     compileSdk 33
 
     defaultConfig {
@@ -44,43 +42,24 @@ dependencies {
     implementation composeBom
     androidTestImplementation composeBom
 
-    // Compose - Material Design 3
-    implementation 'androidx.compose.material3:material3'
+    // Compose - Testing
+    androidTestImplementation "androidx.compose.ui:ui-test-junit4"
 
     // Compose - Android Studio Preview support
     implementation 'androidx.compose.ui:ui-tooling-preview'
     debugImplementation 'androidx.compose.ui:ui-tooling'
 
-    // Compose - Integration with ViewModels
-    implementation 'androidx.lifecycle:lifecycle-viewmodel-compose:2.5.1'
-
-    // Compose - Testing
-    androidTestImplementation "androidx.compose.ui:ui-test-junit4:$compose_version"
-
     // Testing
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
-    testImplementation "org.mockito:mockito-core:5.2.0"
-    testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6"
-
-    // Guava
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-guava:1.4.1'
 
     // CameraX
-    def camerax_version = "1.3.0-alpha05"
+    def camerax_version = "1.1.0-beta01"
     implementation "androidx.camera:camera-core:${camerax_version}"
     implementation "androidx.camera:camera-view:${camerax_version}"
 
-    // Hilt
-    implementation "com.google.dagger:hilt-android:2.44"
-    kapt "com.google.dagger:hilt-compiler:2.44"
-
-    implementation(project(":domain:camera"))
-    implementation(project(':camera-viewfinder-compose'))
-}
-
-// Allow references to generated code
-kapt {
-    correctErrorTypes true
+    // AndroidX Core
+    def core_version = "1.9.0"
+    implementation "androidx.core:core:{$core_version}"
 }
