+++ camera-viewfinder-compose/src/main/AndroidManifest.xml
