+++ settings.gradle
@@ -16,4 +16,4 @@ rootProject.name = "Jetpack Camera"
 include ':app'
 include ':feature:preview'
 include ':domain:camera'
-include ':camerax'
+include ':camera-viewfinder-compose'
